feat: Remove Antelope, io module stuff
This removes anything broken, and shims up the obspy io part to
something that works well enough for testing the HASH algorithm.
Ideally, the FPFIT weird CA file formats will get done, only to verify
the in/out files that ship with the software; otherwise, those may get
axed too.

This is the first commit in a series to rip out the superflous parts of
the lib wrapper, like graphing, or integration with other programs. This
should be a simple lib wrapper used as a dep by other software that does
those things.